requirements.txt
google-generativeai
playwright
speechrecognition
pyttsx3